Weiter gekommen, Name müssen noch Variablen werden
